Moving Scratchpad out of common components
