+++ graylog2-web-interface/src/components/common/index.tsx
@@ -70,7 +70,6 @@ export { default as Pluralize } from './Pluralize';
 export { default as ReactGridContainer } from './ReactGridContainer';
 export { default as ReadOnlyFormGroup } from './ReadOnlyFormGroup';
 export { default as QueryHelper } from './QueryHelper';
-export { default as Scratchpad } from './Scratchpad';
 export { default as ScrollButton } from './ScrollButton';
 export { default as SearchForm } from './SearchForm';
 export { default as Select } from './Select';
