+++ graylog2-web-interface/src/components/scratchpad/index.tsx
